LayoutTests/fast/viewport/viewport-55-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "yes" for key "width" not recognized. Content ignored.
CONSOLE MESSAGE: line 3: Viewport argument value "no" for key "height" not recognized. Content ignored.
viewport size 64x70 scale 5.000000 with limits [5.000000, 5.000000]
